Make charts table constructor more generic, so we can use it in other places (not only zones).
