+++ src/components/pages/serverzones/limitreq.jsx
@@ -6,7 +6,7 @@
  */
 
 import React from 'react';
-import { LimitConnReqConstructor } from './constructors.jsx';
+import ChartsTable from '../../charts-table/index.jsx';
 import Chart from '../../chart/index.jsx';
 import styles from '../../table/style.css';
 
@@ -26,7 +26,7 @@ export const Labels = new Map([
 	['rejected_dry_run', 'Rejected (dry_run)']
 ]);
 
-export default class LimitReq extends LimitConnReqConstructor {
+export default class LimitReq extends ChartsTable {
 	getTitle(){
 		return 'Limit Req';
 	}
@@ -49,7 +49,7 @@ export default class LimitReq extends LimitConnReqConstructor {
 		const { activeCharts } = this.state;
 
 		return Array.from(this.props.data).map(
-			([name, { zone, history }]) => {
+			([name, { obj: zone, history }]) => {
 				let cn = styles['chart-container'];
 				let title = 'Click to view rate graph';
 				let chart = null;
